Added function that get detail from personality type.
